# expression: implement vectorized evaluation for `builtinLocateBinary2ArgsSig`