--- a/expression/builtin_string_vec.go
+++ b/expression/builtin_string_vec.go
@@ -1103,11 +1103,41 @@ func (b *builtinStrcmpSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) 
 }
 
 func (b *builtinLocateBinary2ArgsSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinLocateBinary2ArgsSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	buf0, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf0)
+	buf1, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	if err := b.args[0].VecEvalString(b.ctx, input, buf0); err != nil {
+		return err
+	}
+	if err := b.args[1].VecEvalString(b.ctx, input, buf1); err != nil {
+		return err
+	}
+	result.ResizeInt64(n, false)
+	result.MergeNulls(buf0, buf1)
+	i64s := result.Int64s()
+	for i := 0; i < n; i++ {
+		if result.IsNull(i) {
+			continue
+		}
+		subStr := buf0.GetString(i)
+		if len(subStr) == 0 {
+			i64s[i] = 1
+			continue
+		}
+		i64s[i] = int64(strings.Index(buf1.GetString(i), subStr) + 1)
+	}
+	return nil
 }
 
 func (b *builtinLocateBinary3ArgsSig) vectorized() bool {