Signed-off-by: AerysNan <aerysnan@gmail.com>

PCP #12106

### What problem does this PR solve? <!--add issue link with summary if exists-->

Implement vectorized evaluation for builtinLocateBinary2ArgsSig.

### What is changed and how it works?

```
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-VecBuiltinFunc-4                  34615             31806 ns/op           34288 B/op          5 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-NonVecBuiltinFunc-4               25118             47393 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-VecBuiltinFunc#01-4               39246             30573 ns/op           34288 B/op          5 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-NonVecBuiltinFunc#01-4            24638             45689 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-VecBuiltinFunc#02-4               40785             29424 ns/op           34288 B/op          5 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-NonVecBuiltinFunc#02-4            23210             45961 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-VecBuiltinFunc#03-4               40768             27780 ns/op           34288 B/op          5 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-NonVecBuiltinFunc#03-4            28148             42743 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-VecBuiltinFunc#04-4               42098             29022 ns/op           34288 B/op          5 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-NonVecBuiltinFunc#04-4            25963             45841 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-VecBuiltinFunc#05-4               43052             27083 ns/op           34288 B/op          5 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLocateBinary2ArgsSig-NonVecBuiltinFunc#05-4            28363             42488 ns/op               0 B/op          0 allocs/op 
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test